pyproject.toml
[build-system]
requires = ["poetry==0.12.17"]
build-backend = "poetry.masonry.api"


[tool.nitpick]
style = "https://raw.githubusercontent.com/wemake-services/wemake-python-styleguide/master/styles/nitpick-style-wemake.toml"


[tool.poetry]
name = "resumemakr"
description = "ResumeMakr"
version = "0.1.0"
readme = "README.md"
authors = []


[tool.poetry.dependencies]
python = "3.7.4"
django = "^2.2"
django-split-settings = "^1.0"
django-axes = "^5.0"
django-csp = "^3.5"
django-health-check = "^3.11"
django-http-referrer-policy = "^1.0"
django-feature-policy = "^3.0"
psycopg2 = "^2.8"
gunicorn = "^19.9"
python-decouple = "^3.1"
dump-env = "^1.1"
bcrypt = "^3.1"
ulid2 = "^0.2.0"
graphene-django = "^2.6"


[tool.poetry.dev-dependencies]
django-debug-toolbar = "^2.0"
django-querycount = "^0.7"
nplusone = "^1.0"

flake8-pytest = "^1.3"
flake8-pytest-style = "^0.1.3"
flake8-django = "^0.0.4"
xenon = "^0.5"
nitpick = "^0.21.1"

pytest = "^5.2"
pytest-django = "^3.6"
pytest-cov = "^2.8"
pytest-randomly = "^3.1"
pytest-deadfixtures = "^2.1"
pytest-testmon = "^0.9"
pytest-timeout = "^1.3"

mypy = "^0.720"
django-stubs = "^1.1"

ipython = "^7.8"

sphinx = "^2.2"
sphinx-autodoc-typehints = "^1.8"
doc8 = "^0.8"

yamllint = "^1.18"
safety = "^1.8"
dotenv-linter = "^0.1"
polint = "^0.4"
dennis = "^0.9"
black = {version = "^18.3-alpha.0", allows-prereleases = true}
pytest-watch = "^4.2"
